move models out of api_log_management
