+++ src/actinia_core/rest/api_log_management.py
@@ -29,12 +29,12 @@ from flask import jsonify, make_response
 from flask_restful import Resource
 
 from flask_restful_swagger_2 import swagger
-from flask_restful_swagger_2 import Schema
 
 from actinia_core.common.app import auth
 from actinia_core.common.api_logger import ApiLogger
 from actinia_core.common.api_logger import log_api_call
-from actinia_core.common.response_models import SimpleResponseModel
+from actinia_core.models.openapi.api_log_management import ApiLogListModel
+from actinia_core.models.response_models import SimpleResponseModel
 from actinia_core.rest.user_auth import check_user_permissions
 
 __license__ = "GPLv3"
@@ -42,59 +42,6 @@ __author__ = "Sören Gebbert"
 __copyright__ = "Copyright 2016-2018, Sören Gebbert and mundialis GmbH & Co. KG"
 __maintainer__ = "Sören Gebbert"
 __email__ = "soerengebbert@googlemail.com"
-
-
-class ApiLogEntryModel(Schema):
-    """Response schema for a single API log entry that is used to track all API
-    calls of a user.
-    """
-    type = 'object'
-    properties = {
-        'time_stamp': {
-            'type': 'string',
-            'description': 'The time stamp of the API call'
-        },
-        'node': {
-            'type': 'string',
-            'description': 'The node that executed the API call'
-        },
-        'endpoint': {
-            'type': 'string',
-            'description': 'The endpoint of the API call'
-        },
-        'method': {
-            'type': 'string',
-            'description': 'The HTTP method of the request'
-        },
-        'path': {
-            'type': 'string',
-            'description': 'The path of the REST API call'
-        },
-        'url': {
-            'type': 'string',
-            'description': 'The request URL'
-        },
-        'request_str': {
-            'type': 'string',
-            'description': 'The request string'
-        }
-    }
-    required = ["time_stamp", "node", "endpoint",
-                "method", "path", "url", "request_str"]
-
-
-class ApiLogListModel(Schema):
-    """Response schema that represents a list of API log entries.
-    """
-    type = 'object'
-    properties = {
-        'api_log_list': {
-            'type': 'array',
-            'items': ApiLogEntryModel,
-            'description': 'A list of ApiLogEntryModel objects'
-        }
-    }
-    required = ["api_log_list"]
 
 
 class APILogResource(Resource):
